requirements.txt
discord.py
boxsdk[jwt]